Removed erroneous "FROM"; Changed some tags to lowercase.
A single year as a date was sometimes displayed as "FROM" which I felt was incorrect.
I decided to change months from MAR to Mar and similar date expressions from ABT to Abt. to be more friendly in my app. It seemed to change them at the source as they were parsed was the easiest/most effective way.

There also seemed to be a bug where  the period parsing was mapping ABT to Estimate rather than About. I changed that.